% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sampling.R
\name{calc_strat_samples}
\alias{calc_strat_samples}
\title{Calculate stratified sample}
\usage{
calc_strat_samples(data, grp_by, conf_level, margin_of_error, prop = 0.5)
}
\arguments{
\item{data}{the dataframe you want to sample}

\item{grp_by}{the group or strata}

\item{conf_level}{the confidence level}

\item{margin_of_error}{the margin of error}

\item{prop}{the proportion, default = 0.5}
}
\description{
A function that allows you sample a population based on strata. It will output the \emph{sampled} population, meaning the returned data frame will be your sample output.
}
\examples{
output <- calc_strat_samples(data =  DSA_VIII_data, grp_by = Neighbourhood, conf_level = 0.95, margin_of_error = 0.05)
}
