% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_loa.R
\name{create_loa}
\alias{create_loa}
\title{Automatically produce a LOA from a Kobo Tool and survey data}
\usage{
create_loa(
  kobo_survey = NULL,
  data = NULL,
  level = 0.95,
  group_vars = character(0),
  sm_sep = "/",
  print_class_type = TRUE
)
}
\arguments{
\item{kobo_survey}{Your Kobo survey. Must have columns type and name. Leave blank if you want the LOA from the dataframe instead}

\item{data}{Your survey data, for capturing composite indicators not from your tool. Leave blank if only tool is required.}

\item{level}{confidence interval, defaults to 0.95}

\item{group_vars}{disaggregation variables, to be provided as a string. E.g. c('admin1', 'admin2). Leave blank for no disaggregation.}

\item{sm_sep}{defaults to '/'}

\item{print_class_type}{TRUE/FALSE whether to output any analysis_type assumed based on the data type.}
}
\value{
A dataframe containing
\describe{
\item{analysis_var}{name of the indicator}
\item{analysis_type}{analysis to perform}
\item{level}{Confidence Interval}
\item{group_var}{indicators to disaggregate by}
}
}
\description{
This function produces a LOA to be used in the production of results tables from
\code{analysistools::create_analysis}.
The function creates a LOA based on the Kobo server tool, inferring the analysis_type from
the 'type' column in the tool.
It adds to this by looking at any data in the survey that is not present in the tool
and attempts to guess based on the class of the column.
}
\examples{
create_loa(kobo_survey = tool, data = clean_data, group_vars = c('admin_1', 'admin2'))

}
