% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variable_tracker.R
\name{create_variable_tracker}
\alias{create_variable_tracker}
\title{Create a Variable Tracker to Compare Raw and Cleaned Datasets}
\usage{
create_variable_tracker(raw_data, clean_data)
}
\arguments{
\item{raw_data}{A data frame representing the original/raw dataset before cleaning.}

\item{clean_data}{A data frame representing the cleaned dataset.}
}
\value{
A data frame with three columns:
\describe{
\item{Variable}{The name of the variable that was added or removed.}
\item{Action}{Indicates whether the variable was "added" (present in \code{clean_data} but not in \code{raw_data})
or "remove" (present in \code{raw_data} but not in \code{clean_data}).}
\item{Rationale}{A blank field where the rationale for each action can be added manually.}
}
}
\description{
This function compares the column names of a raw dataset and a cleaned dataset
and generates a tracker showing which variables have been added or removed.
It should be included as a tab in the data submission to HQ.
}
\examples{
library(dplyr)
raw_data <- data.frame(a = 1:3, b = 4:6, c = 7:9)
clean_data <- data.frame(a = 1:3, d = 10:12)
create_variable_tracker(raw_data, clean_data)

}
