% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kobo_metadata.R
\name{get_kobo_metadata}
\alias{get_kobo_metadata}
\title{Get Kobo Metadata Duration}
\usage{
get_kobo_metadata(dataset, asset_id, kobo_token = NULL, remove_geo = TRUE)
}
\arguments{
\item{dataset}{the name of your survey dataframe.}

\item{asset_id}{the id of your kobo object. Can be taken from your survey URL.}

\item{kobo_token}{your kobo API key. Optional in case it is already set as a system variable.}

\item{remove_geo}{whether to remove time taken to get gps coordinates from the summary. Defaults to TRUE.}
}
\description{
Adds the duration of each interview as a column to your dataset based on the metadata logs
}
\examples{
get_kobo_metadata(dataset = data_in_processing, asset_id = "antAdT3siLrnjTdfcdYcFY")
}
