% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clog_check.R
\name{review_cleaning_log}
\alias{review_cleaning_log}
\title{Run basic checks to ensure cleaning log validity}
\usage{
review_cleaning_log(
  tool_path,
  survey_sheet = "survey",
  choice_sheet = "choices",
  sample_survey_data_path,
  sample_survey_data_path_sheet = NULL,
  file_path,
  file_path_sheet = "cleaning_log",
  output_dir,
  output_as_list = TRUE,
  output_as_excel = TRUE
)
}
\arguments{
\item{tool_path}{the path to the Kobo tool file}

\item{survey_sheet}{the sheet name, defaults to "survey"}

\item{choice_sheet}{the choices sheet, defaults to "choices"}

\item{sample_survey_data_path}{a filepath to some sample results data. This is just used to identify valid column headers.}

\item{sample_survey_data_path_sheet}{the sheet name, defaults to NULL (selects the first sheet in the excel)}

\item{file_path}{the filepath to the clog}

\item{file_path_sheet}{the sheetname for the clog, defaults to "cleaning_log"}

\item{output_dir}{the location to export the cleaning_log_report to}

\item{output_as_list}{defaults to TRUE}

\item{output_as_excel}{defaults to TRUE}
}
\description{
A function runs two basic checks on a cleaning log. The first check is check whether the change_type is consistent with the new value and old value. The second is to see whether select one and select multiple new values are valid based on tool entry.

Please note that the separator used for select multiple questions should be "/".
}
\examples{
process_cleaning_log(
  tool_path = r"(01_Tool/REACH_2024_SOM_DSA_Survey_Tool_VIII.xlsx)",
  sample_survey_data_path = r"(DSA_2024_REACH_SOM_Processed_Data_2024-12-24.xlsx)",
  file_path = r"(01CLeaning scripts\_code\outputs\kobo_tool.xlsx)",
  output_dir = r"(Downloads)",
  output_as_list = TRUE,
  output_as_excel = FALSE
 )
}
