% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ki_database.R
\name{update_ki_database}
\alias{update_ki_database}
\title{Update Key Informant (KI) Database}
\usage{
update_ki_database(
  data,
  date,
  district,
  region,
  idp_site,
  ki_name,
  ki_age,
  ki_role,
  ki_status,
  ki_contact,
  FieldOfficer,
  r_c
)
}
\arguments{
\item{data}{A data frame containing the new data to be added.}

\item{date}{Column name (string) for the date of data collection.}

\item{district}{Column name (string) for the district.}

\item{region}{Column name (string) for the region.}

\item{idp_site}{Column name (string) for the IDP site.}

\item{ki_name}{Column name (string) for the Key Informant's name.}

\item{ki_age}{Column name (string) for the Key Informant's age.}

\item{ki_role}{Column name (string) for the Key Informant's role.}

\item{ki_status}{Column name (string) for the Key Informant's status.}

\item{ki_contact}{Column name (string) for the Key Informant's contact information.}

\item{FieldOfficer}{Column name (string) for the responsible field officer.}

\item{r_c}{A string to categorize the research category.}
}
\value{
An updated data frame combining the original KI database and new data.
}
\description{
This function appends new data to an existing Key Informant (KI) database, formats the data, and saves it as an Excel file.
The function includes the file path for the KI database.
The function includes parameters for date, district, region, IDP site, name, age, role, status, contact information, and field officer.
It should be run for all RCs that use KI data.
The parameters for the function are used to rename the columns returned by the survey into the standard data model for the KI database, making it possible to append data from different RCs.
}
\examples{
\dontrun{

update_ki_database(
  data = df,
  date = "today",
  district = "district_name",
  region = "localisation_region_label",
  idp_site = "idp_site",
  ki_name = "ki_name",
  ki_age = "ki_age",
  ki_role = "ki_role",
  ki_status = "ki_status",
  ki_contact = "ki_contact",
  FieldOfficer = "Responsible_FO",
  r_c = "DSA"
}
}
