% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_kobo_api.R
\name{get_kobo_data}
\alias{get_kobo_data}
\title{Kobo Server Query Function}
\usage{
get_kobo_data(asset_id, un)
}
\arguments{
\item{asset_id}{Character. The asset ID of the survey.}

\item{un}{Character. Your Kobo username.}
}
\value{
A data frame of the Kobo survey results.
}
\description{
Retrieve Kobo data from the server by providing the asset ID.
}
