% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_kobo_api.R
\name{get_kobo_data}
\alias{get_kobo_data}
\title{Kobo Server Query function
A function that allows you to retrieve kobo data from the server by providing the asset ID}
\usage{
get_kobo_data(asset_id)
}
\arguments{
\item{asset_id}{What is the asset ID of the survey youre interested in?}
}
\description{
Kobo Server Query function
A function that allows you to retrieve kobo data from the server by providing the asset ID
}
\examples{
#kobo_data <- get_kobo_data(asset_id = "a38DR4AhAH2rDqP6e5YHLm") # TEST SURVEY
}
\keyword{asset}
