% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kobo_metadata.R
\name{get_kobo_metadata}
\alias{get_kobo_metadata}
\title{Get Kobo Interview Duration}
\usage{
get_kobo_metadata(dataset, remove_geo = TRUE)
}
\arguments{
\item{dataset}{A data frame containing your Kobo survey data.}

\item{remove_geo}{Logical. If TRUE (default), durations related to GPS capture are excluded.}
}
\value{
A list with:
\describe{
\item{df_and_duration}{The original dataset with added interview duration and metadata start time.}
\item{audit_files_length}{The parsed audit log with duration calculations.}
}
}
\description{
Calculates interview duration using KoboToolbox metadata logs and merges
this information with your dataset.
}
\examples{
\dontrun{
get_kobo_metadata(dataset = data_in_processing, asset_id = "abc123XYZ", un = "my_kobo_username")
}
}
